clean: Move pytests
